Moved functions from posts.js -> utility.js, tidied code
